refactor(trusty-search): split xlsx.rs under the 500-SLOC cap, cover the post-read size check (#4894)
CI's line-cap gate refused xlsx.rs at 611 SLOC. check_line_cap.sh counts an
inline `#[cfg(test)]` module against the 500-SLOC PRODUCTION cap, and this
file's three fixture builders put nearly all of that bulk in tests. A
basename of exactly `tests.rs` is classified as a test file (3000-SLOC cap),
so xlsx.rs becomes xlsx/mod.rs (99 SLOC) + xlsx/tests.rs (541 SLOC). A child
module already sees the parent's private items, so no production code moved
and nothing needed its visibility widened.

Two tests cover branches that had none. `read_package_bytes` re-checks the
length on the bytes actually read rather than trusting `metadata`, and that
branch was unreachable from the suite: `test_oversized_package_reports_too_large`
writes cap+1 bytes, which the earlier metadata check catches, so execution
never arrived. Deleting the branch outright, and flipping its `>` to `>=`
(which would reject a legal workbook sitting exactly on the cap), both left
17/17 green.

- `test_document_exactly_at_size_cap_is_accepted` pins the boundary.
- `test_stream_longer_than_its_stat_is_still_refused` feeds a FIFO, whose
  metadata reports length 0, then writes cap+1 bytes through it — the
  "grew after we stat'd it" case the branch exists for, made deterministic.

Each mutant now kills exactly one test, and skipping just these two restores
the old 17/17 pass under both.

MAX_WORKBOOK_ENTRIES 4096 -> 16384. A chart-dense dashboard emits ~4-5 parts
per chart, so 4096 would refuse a real ~850-chart workbook, and a too-tight
cap fails legitimate files. The bound that matters is metadata cost at the
container ceiling (~115k entries), which 16384 stays well inside.
